+++ src/main/java/org/springframework/social/showcase/repository/UserConnectionRepository.java
@@ -10,7 +10,6 @@ import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.data.jpa.repository.Query;
 import org.springframework.social.showcase.model.ConnectionProperty;
 import org.springframework.social.showcase.model.UserConnection;
-import org.springframework.stereotype.Repository;
 
 public interface UserConnectionRepository extends JpaRepository<UserConnection, Long> {
 
@@ -23,26 +22,26 @@ public interface UserConnectionRepository extends JpaRepository<UserConnection, 
 	@Query("select u.hashId from UserConnection uc left join uc.user u where uc.providerId = ?1 and uc.providerUserId = ?2")
 	public List<String> findUserHashIdsByProviderIdAndProviderUserId(String providerId, String providerUserId);
 	
-	public List<ConnectionProperty> findByUserIdAndProviderIdOrderByRankAsc(Integer id, String providerId);
+	public List<ConnectionProperty> findByUserIdAndProviderIdOrderByRankAsc(Long id, String providerId);
 	
-	public Optional<ConnectionProperty> findTopByUserIdAndProviderIdOrderByRankAsc(Integer id, String providerId);
+	public Optional<ConnectionProperty> findTopByUserIdAndProviderIdOrderByRankAsc(Long id, String providerId);
 	
-	public List<ConnectionProperty> findByUserIdOrderByProviderIdAscRankAsc(Integer id);
+	public List<ConnectionProperty> findByUserIdOrderByProviderIdAscRankAsc(Long id);
 	
-	public Optional<UserConnection> findByUserIdAndProviderIdAndProviderUserId(Integer id, String providerId, String providerUserId);
+	public Optional<UserConnection> findByUserIdAndProviderIdAndProviderUserId(Long id, String providerId, String providerUserId);
 	
 	@Query("select coalesce(max(uc.rank) + 1, 1) as rank from UserConnection uc where uc.user.id = ?1 and uc.providerId = ?2")
-	public Integer findNextRankByUserIdAndProviderId(Integer id, String providerId);
+	public Integer findNextRankByUserIdAndProviderId(Long id, String providerId);
 		
 //	@Query("delete from UserConnection uc where uc.user.id = ?1 and uc.providerId = ?2")
 //	@Modifying
 	@Transactional
-	public void deleteByUserIdAndProviderId(Integer id, String providerId);	
+	public void deleteByUserIdAndProviderId(Long id, String providerId);	
 //	public void deleteByProviderId(String providerId);
 	
 //	@Modifying
 	@Transactional
-	public Integer deleteByUserIdAndProviderIdAndProviderUserId(Integer id, String providerId, String providerUserId);
+	public Integer deleteByUserIdAndProviderIdAndProviderUserId(Long id, String providerId, String providerUserId);
 
 }
